requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
joblib==1.4.2
numpy==2.1.2
packaging==24.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
